Refactors authentication
