+++ src/app/auth.js
@@ -1,7 +1,15 @@
-import hash from "../helpers/hash";
+import hash from "../utils/hash";
+import history from "../utils/history";
 
-const token = () => {
-  return hash.access_token;
+const isAuthenticated = () => {
+  const token = hash.access_token || localStorage.getItem("token");
+  if (token) {
+    localStorage.setItem("token", token);
+    history.replace("/");
+    return token;
+  } else {
+    history.replace("/login");
+  }
 };
 
-export default token();
+export default isAuthenticated();
