+++ src/utils/hash.js
@@ -9,4 +9,6 @@ const hash = window.location.hash
     return initial;
   }, {});
 
+window.location.hash = "";
+
 export default hash;
